WIP P1: swap emulator core to ARMSX2 iOSv2.3.0 (clean, un-hacked)
First step of the ARMSX2 rebase (Option A). Replaces our bloated core
(Counters 7179->1073, iR5900 7135->2738, recVTLB 2253->1072, DarwinMisc
2172->1265) with iOSv2.3.0's clean shipping-iOS versions. 3rdparty kept
(near-identical). EXPECTED RED: our old ios_main.mm/CMake still reference the
previous core — the CI failure maps the reconciliation surface for the next
steps (adopt their native layer + CMake, then re-skin as ELORIS-PRISM).

Installable build stays #101 until the migration goes green + device-tested.

+++ src/cpp/pcsx2/GS/GSCapture.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2002-2025 PCSX2 Dev Team
+// SPDX-FileCopyrightText: 2002-2026 PCSX2 Dev Team
 // SPDX-License-Identifier: GPL-3.0+
 
 #pragma once
@@ -21,7 +21,7 @@ namespace GSCapture
 {
 	bool BeginCapture(float fps, GSVector2i recommendedResolution, float aspect, std::string filename);
 	bool DeliverVideoFrame(GSTexture* stex);
-	void DeliverAudioPacket(const s16* frames); // AudioStream::CHUNK_SIZE
+	void DeliverAudioPacket(const float* frames); // AudioStream::CHUNK_SIZE
 	void EndCapture();
 
 	bool IsCapturing();
@@ -38,7 +38,7 @@ namespace GSCapture
 	CodecList GetVideoCodecList(const char* container);
 	CodecList GetAudioCodecList(const char* container);
 
-	using FormatName = std::pair<int , std::string>; // id,name
+	using FormatName = std::pair<int, std::string>; // id,name
 	using FormatList = std::vector<FormatName>;
 	FormatList GetVideoFormatList(const char* codec);
 }; // namespace GSCapture
