+++ src/cpp/pcsx2/IPU/IPU.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2002-2025 PCSX2 Dev Team
+// SPDX-FileCopyrightText: 2002-2026 PCSX2 Dev Team
 // SPDX-License-Identifier: GPL-3.0+
 
 #pragma once
@@ -284,7 +284,7 @@ extern bool FMVstarted;
 extern bool EnableFMV;
 
 alignas(16) extern tIPU_cmd ipu_cmd;
-extern uint eecount_on_last_vdec;
+extern u64 eecount_on_last_vdec;
 
 extern void ipuReset();
