+++ src/cpp/pcsx2/SPU2/spu2freeze.cpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2002-2025 PCSX2 Dev Team
+// SPDX-FileCopyrightText: 2002-2026 PCSX2 Dev Team
 // SPDX-License-Identifier: GPL-3.0+
 
 #include "SPU2/defs.h"
@@ -34,7 +34,7 @@ struct SPU2Savestate::DataBlock
 	u16 OutPos;
 	u16 InputPos;
 	u32 Cycles;
-	u32 lClocks;
+	u64 lClocks;
 	int PlayMode;
 };
